Remove redundant comments.
/* */ comments are a little more preferred in C, unlike the C++ // comments.
Although both work since C99.

Put comments on top of code blocks to show why or what each block is doing.
You don't have to explain every step; programmers are pretty smart :)

http://blog.codinghorror.com/code-tells-you-how-comments-tell-you-why/

For example:
    combinationArray[total] += 1; // Increment how many times that number has been rolled

"+= 1" *means* increment; do we need to write a comment to tell people that?

---
note: Most of the comments in this revised code are even *too* much.
"zero out the array" and "print out the array" are pretty obvious.
---